pyproject.toml
[project]
name = "refactoring-benchmark"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = false

dependencies = [
    "podman (>=5.6.0,<6.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "scipy (>=1.16.3,<2.0.0)",
    "tqdm (>=4.66.0,<5.0.0)",
    "matplotlib (>=3.8.0,<4.0.0)",
    "anthropic (>=0.40.0,<1.0.0)",
    "joblib (>=1.5.3,<2.0.0)",
    "unidiff (>=0.7.5,<0.8.0)",
    "whatthepatch (>=1.0.7,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)",
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-cov (>=4.0.0)",
    "ruff (>=0.15.0,<0.16.0)",
    "isort (>=7.0.0,<8.0.0)"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=refactoring_benchmark",
    "--cov-report=term-missing",
    "--cov-report=html"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests requiring Docker",
    "unit: marks tests as unit tests (no external dependencies)"
]

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''